Warm_up_Challenges/Counting_Valleys/Solution_py/Solution.py
#!/bin/python3

import math
import os
import random
import re
import sys

#
# Complete the 'countingValleys' function below.
#
# The function is expected to return an INTEGER.
# The function accepts following parameters:
#  1. INTEGER steps
#  2. STRING path
#


def countingValleys(steps, path):
    # Write your code here
    M = 0
    step = 0
    route = []
    for i in range(steps):
        if path[i] == 'U':
            step += 1
        else:
            step -= 1
        route.append(step)
        if step == 0 and route[-2] < 0:
            M += 1
    return M


if __name__ == '__main__':
    fptr = open(os.environ['OUTPUT_PATH'], 'w')

    steps = int(input().strip())

    path = input()

    result = countingValleys(steps, path)

    fptr.write(str(result) + '\n')

    fptr.close()
